+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## UNRELEASED
 ### Added
+* ui: Add UI for deposit return vouchers (APPS-1643)
 ### Changed
 ### Removed
 * ui/core: Remove everything related to the old deposit return voucher feature